Refactoring Fury Serialization Integration
